Rename SelectConditions component file and adjust imports
