+++ designer/client/link-create.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SelectConditions from "./conditions/select-conditions";
+import SelectConditions from "./conditions/SelectConditions";
 import { ErrorMessage } from "@govuk-jsx/error-message";
 import { clone } from "@xgovformbuilder/model";
 import classNames from "classnames";
