+++ designer/client/link-edit.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SelectConditions from "./conditions/select-conditions";
+import SelectConditions from "./conditions/SelectConditions";
 import { clone } from "@xgovformbuilder/model";
 
 import { DataContext } from "./context";
